+++ common/src/test/scala/dx/util/PosixPathTest.scala
@@ -8,6 +8,7 @@ class PosixPathTest extends AnyFlatSpec with Matchers {
     val path = PosixPath("/a/b/c")
     path.toString shouldBe "/a/b/c"
     path.parts shouldBe Vector("a", "b", "c")
+    path.nameCount shouldBe 3
     path.isAbsolute shouldBe true
     path.getName shouldBe Some("c")
     val parent = PosixPath(Vector("a", "b"), isAbsolute = true)
@@ -28,10 +29,21 @@ class PosixPathTest extends AnyFlatSpec with Matchers {
     }
   }
 
+  it should "parse the root path" in {
+    val path = PosixPath("/")
+    path.toString shouldBe "/"
+    path.parts shouldBe empty
+    path.nameCount shouldBe 0
+    path.isAbsolute shouldBe true
+    path.getName shouldBe None
+    path.getParent shouldBe None
+  }
+
   it should "parse a relative path" in {
     val path = PosixPath("a/b/c")
     path.toString shouldBe "a/b/c"
     path.parts shouldBe Vector("a", "b", "c")
+    path.nameCount shouldBe 3
     path.isAbsolute shouldBe false
     path.getName shouldBe Some("c")
     val parent = PosixPath(Vector("a", "b"), isAbsolute = false)